+++ CHANGELOG.md
@@ -5,6 +5,7 @@ Get upgrade notes from Sprockets 3.x to 4.x at https://github.com/rails/sprocket
 ## Master
 
 - Remove remaining support for Ruby < 2.4.[#672](https://github.com/rails/sprockets/pull/672)
+- Find all asset by a regexp filemask or a filepath.[#698](https://github.com/rails/sprockets/pull/698)
 
 ## 4.0.2